// src/react-app/lib/finance-types.ts
export interface CategoryGroup {
  id: string;
  name: string;
}

export interface Category {
  id: string;
  name: string;
  type: "receita" | "despesa";
  groupId: string;
  isDefault?: boolean;
}
// ...
export const CATEGORY_GROUPS: CategoryGroup[] = [
  { id: "grp-1", name: "Receitas Operacionais" },
  { id: "grp-2", name: "Custos Operacionais" },
  { id: "grp-3", name: "Despesas Operacionais e Outras Receitas" },
  { id: "grp-4", name: "Atividades de Investimento" },
  { id: "grp-5", name: "Atividades de Financiamento" },
];